node js. BD

+++ routes/upload.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const formidable = require('formidable');
 const fs = require('fs');
 const path = require('path');
+const mongoose = require('mongoose');
 const config = require('../config.json');
 
 router.get('/', function (req, res) {
@@ -19,13 +20,33 @@ router.post('/', function (req, res) {
         if (err) {
             return res.json({status: 'Не удалось загрузить картинку'});
         }
-        fs.rename(files.photo.path, path.join(config.upload, files.photo.name), function (err) {
-            if (err) {
-                fs.unlink(path.join(config.upload, files.photo.name));
-                fs.rename(files.photo.path, files.photo.name);
-            }
-            res.json({status: 'Картинка успешно загружена'});
-        });
+        if (!fields.name) {
+            return res.json({status: 'Не указано описание картинки!'});
+        }
+        const Model = mongoose.model('pic');
+
+        fs
+            .rename(files.photo.path, path.join(config.upload, files.photo.name), function (err) {
+                if (err) {
+                    fs.unlink(path.join(config.upload, files.photo.name));
+                    fs.rename(files.photo.path, files.photo.name);
+                }
+                let dir = config.upload.substr(config.upload.indexOf('/'));
+                // const item = new Model({name: fields.name, picture: path.join(dir, files.photo.name)});
+                // item.save().then(pic => {
+                //   res.json({status: 'Картинка успешно загружена'});
+                // });
+
+                const item = new Model({name: fields.name});
+                item.save().then(pic => {
+                    Model.update({_id: pic._id}, {$set: {picture: path.join(dir, files.photo.name)}}, {upsert: true})
+                    .then(
+                        i => res.json({status: 'Картинка успешно загружена'}),
+                    e => res.json({status: e.message})
+                );
+            });
+
+            });
     });
 });
